+++ src/ast/transformation.test.ts
@@ -25,5 +25,47 @@ await doSomething();`;
 
       expect(() => transform(code, {})).not.toThrow();
     });
+
+    it("should preserve tabs", () => {
+      const code = `\t\tconsole.log("Hello Tabs")`;
+
+      const { code: result } = transform(code, {
+        StringLiteral(path) {
+          path.node.value = `Tabs still here?`;
+        }
+      });
+
+      expect(result).toBe(`\t\tconsole.log("Tabs still here?")`);
+    });
+
+    it("should preserve tabs (multi-lines)", () => {
+      const code = `function doSomethingWithTabs() {
+\tconst message = "Hello!";
+\t\tconsole.log(message);
+}`;
+
+      const { code: result } = transform(code, {
+        StringLiteral(path) {
+          path.node.value = `Tabs still here?`;
+        }
+      });
+
+      expect(result).toBe(`function doSomethingWithTabs() {
+\tconst message = "Tabs still here?";
+\t\tconsole.log(message);
+}`);
+    });
+
+    it("should preserve spaces", () => {
+      const code = `  console.log("Hello Tabs")`;
+
+      const { code: result } = transform(code, {
+        StringLiteral(path) {
+          path.node.value = `Tabs still here?`;
+        }
+      });
+
+      expect(result).toBe(`  console.log("Tabs still here?")`);
+    });
   });
 });